requirements.txt
click>=8.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiosqlite>=0.19.0
rich>=13.0.0
loguru>=0.7.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
